requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.1
fsspec==2024.6.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.0.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sympy==1.13.3
tifffile==2025.5.26
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
